+++ continuous_integration/travis/install.sh
@@ -91,6 +91,7 @@ if [ -n "$NUMPY_VERSION" ]; then
 fi
 
 if [[ "$COVERAGE" == "true" ]]; then
+    # TODO: unpin when https://github.com/nedbat/coveragepy/issues/883 is fixed
     # Weird issues with recent version of coverage: unpin when not causing
     # pytest to raise INTERNALERROR exceptions.
     PIP_INSTALL_PACKAGES="$PIP_INSTALL_PACKAGES coverage==4.5.4 pytest-cov codecov"